multi: add new bins module
