+++ src/wallet/mod.rs
@@ -62,7 +62,6 @@ mod cusf_block_producer;
 pub mod error;
 mod mine;
 mod mnemonic;
-pub mod signet_miner;
 
 type BundleProposals = Vec<(M6id, BlindedM6<'static>, Option<PendingM6idInfo>)>;
